+++ core/alarm/src/main/java/com/yapp/alarm/receivers/AlarmReceiver.kt
@@ -13,7 +13,7 @@ import com.yapp.analytics.AnalyticsEvent
 import com.yapp.analytics.AnalyticsHelper
 import com.yapp.domain.model.Alarm
 import com.yapp.domain.model.toTimeString
-import com.yapp.domain.repository.UserDataRepository
+import com.yapp.domain.repository.FortuneRepository
 import com.yapp.domain.usecase.AlarmUseCase
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.CoroutineScope
@@ -34,7 +34,7 @@ class AlarmReceiver : BroadcastReceiver() {
     lateinit var androidAlarmScheduler: AndroidAlarmScheduler
 
     @Inject
-    lateinit var userDataRepository: UserDataRepository
+    lateinit var fortuneRepository: FortuneRepository
 
     @Inject
     lateinit var alarmUseCase: AlarmUseCase
@@ -107,13 +107,13 @@ class AlarmReceiver : BroadcastReceiver() {
                                 ),
                             ),
                         )
-                        val existingId = userDataRepository.firstDismissedAlarmIdFlow.firstOrNull()
+                        val existingId = fortuneRepository.firstDismissedAlarmIdFlow.firstOrNull()
                         if (existingId == null) {
                             // 첫 번째 알람 해제 기록
-                            userDataRepository.saveFirstDismissedAlarmId(alarmId)
+                            fortuneRepository.saveFirstDismissedAlarmId(alarmId)
                         } else if (existingId != alarmId) {
                             // 두 번째 알람 해제 감지 - 기존 기록 삭제
-                            userDataRepository.clearDismissedAlarmId()
+                            fortuneRepository.clearDismissedAlarmId()
                         }
                     }
