+++ data/src/main/java/com/yapp/data/local/datasource/AlarmLocalDataSource.kt
@@ -5,6 +5,8 @@ import com.yapp.domain.model.Alarm
 import kotlinx.coroutines.flow.Flow
 
 interface AlarmLocalDataSource {
+    val firstDismissedAlarmIdFlow: Flow<Long?>
+
     fun getAllAlarms(): Flow<List<Alarm>>
     fun getPagedAlarms(limit: Int, offset: Int): Flow<List<Alarm>>
     fun getAlarmsByTime(hour: Int, minute: Int, isAm: Boolean): Flow<List<Alarm>>
@@ -14,4 +16,6 @@ interface AlarmLocalDataSource {
     suspend fun updateAlarmActive(id: Long, active: Boolean): Int
     suspend fun getAlarm(id: Long): Alarm?
     suspend fun deleteAlarm(id: Long): Int
+    suspend fun saveFirstDismissedAlarmId(alarmId: Long)
+    suspend fun clearDismissedAlarmId()
 }
