+++ data/src/main/java/com/yapp/data/local/datasource/AlarmLocalDataSourceImpl.kt
@@ -3,6 +3,7 @@ package com.yapp.data.local.datasource
 import com.yapp.data.local.AlarmDao
 import com.yapp.data.local.AlarmEntity
 import com.yapp.data.local.toDomain
+import com.yapp.datastore.UserPreferences
 import com.yapp.domain.model.Alarm
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
@@ -10,7 +11,10 @@ import javax.inject.Inject
 
 class AlarmLocalDataSourceImpl @Inject constructor(
     private val alarmDao: AlarmDao,
+    private val userPreferences: UserPreferences,
 ) : AlarmLocalDataSource {
+    override val firstDismissedAlarmIdFlow: Flow<Long?> = userPreferences.firstDismissedAlarmIdFlow
+
     override fun getAllAlarms(): Flow<List<Alarm>> {
         return alarmDao.getAllAlarms()
             .map { alarmEntities -> alarmEntities.map { it.toDomain() } }
@@ -52,5 +56,13 @@ class AlarmLocalDataSourceImpl @Inject constructor(
 
     override suspend fun deleteAlarm(id: Long): Int {
         return alarmDao.deleteAlarm(id)
+    }
+
+    override suspend fun saveFirstDismissedAlarmId(alarmId: Long) {
+        userPreferences.saveFirstDismissedAlarmId(alarmId)
+    }
+
+    override suspend fun clearDismissedAlarmId() {
+        userPreferences.clearDismissedAlarmId()
     }
 }
