+++ data/src/main/java/com/yapp/data/local/datasource/FortuneLocalDataSource.kt
@@ -1,11 +1,8 @@
-package com.yapp.domain.repository
+package com.yapp.data.local.datasource
 
 import kotlinx.coroutines.flow.Flow
 
-interface UserDataRepository {
-    val userIdFlow: Flow<Long?>
-    val userNameFlow: Flow<String?>
-    val onboardingCompletedFlow: Flow<Boolean>
+interface FortuneLocalDataSource {
     val fortuneIdFlow: Flow<Long?>
     val fortuneDateFlow: Flow<String?>
     val fortuneImageIdFlow: Flow<Int?>
@@ -13,15 +10,11 @@ interface UserDataRepository {
     val hasNewFortuneFlow: Flow<Boolean>
     val firstDismissedAlarmIdFlow: Flow<Long?>
 
-    suspend fun saveUserId(userId: Long)
-    suspend fun saveUserName(userName: String)
-    suspend fun setOnboardingCompleted()
     suspend fun saveFortuneId(fortuneId: Long)
     suspend fun markFortuneAsChecked()
     suspend fun saveFortuneImageId(imageResId: Int)
     suspend fun saveFortuneScore(score: Int)
     suspend fun saveFirstDismissedAlarmId(alarmId: Long)
     suspend fun clearDismissedAlarmId()
-    suspend fun clearUserData()
     suspend fun clearFortuneId()
 }
