+++ data/src/main/java/com/yapp/data/local/di/DataSourceModule.kt
@@ -2,6 +2,8 @@ package com.yapp.data.local.di
 
 import com.yapp.data.local.datasource.AlarmLocalDataSource
 import com.yapp.data.local.datasource.AlarmLocalDataSourceImpl
+import com.yapp.data.local.datasource.FortuneLocalDataSource
+import com.yapp.data.local.datasource.FortuneLocalDataSourceImpl
 import com.yapp.data.local.datasource.UserLocalDataSource
 import com.yapp.data.local.datasource.UserLocalDataSourceImpl
 import dagger.Binds
@@ -18,6 +20,12 @@ abstract class DataSourceModule {
     abstract fun bindsAlarmDataSource(
         alarmLocalDataSource: AlarmLocalDataSourceImpl,
     ): AlarmLocalDataSource
+
+    @Binds
+    @Singleton
+    abstract fun bindsFortuneDataSource(
+        fortuneLocalDataSource: FortuneLocalDataSourceImpl,
+    ): FortuneLocalDataSource
 
     @Binds
     @Singleton
